Refactoring: pushed some code to mymodules
I've created and added a module "add-sr.js" which gets and filters
for output issues from search query.
As parameters you need to pass query and a callback function.
The argument of a callback function is object with "searchresults" array
which we can pass further (and we actually do) "res.render" function as
it's second parameter
